requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
aiohttp==3.9.1
sentence-transformers==2.7.0
transformers==4.36.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.1
black==23.11.0
isort==5.12.0
openai==1.3.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
websockets==11.0.3
python-dotenv==1.0.0
httpx==0.25.2
psycopg2-binary==2.9.9
